Separate tests into header failures and read failures.
There are now two sets of tests -- those for which the file is too short
to have a valid header, which should produce an error on `connect`, and
those which have some invalid data within a valid pcap file.  The latter
case should terminate `listen` when a short read occurs.  Any available
segment of the last packet will be lost.

Signed-off-by: Mindy Preston <mindy.preston@docker.com>